final/model2/requirements.txt
grpcio==1.66.2
grpcio-tools==1.66.2
protobuf==5.27.3
numpy==1.26.0
opencv-python==4.9.0.80
easyocr>=1.7.1
Pillow==10.0.0
